Organize backend code, create FastAPI dependency function
